rkt: Move fetch code to images.go
No code or functional changes.
